// src/chrome/api/findnow/inc/Subject/SubjectOptions.ts
export type SubjectOptions = {
    subject: string;
    dirPath: string;
    use_abbreviation: boolean;
    abbreviation: string;
    filenameFormat: number;
    cutSubject: boolean;
    cutFilename: boolean;
    filenames_toascii: boolean;
    use_iso_date: boolean;
    add_time_to_name: boolean;
    pattern: string;
};